refactoring tests to handle expcetions
- testRecaptchaSolve
The test url has a problem and it seems to block the test and need to
be skipped until a new url are provided.

This organization has been flagged for violating hCaptcha's terms of service

- testIncorrectWebsite

this test is created to test correct website and it throw an expcetion
now it has an expectedException.

- testIncorrectMinScore

This test also throw an expection because the Incorrect Min Score
and now ot has an expectedException.

- testIncorrectRecognizingThreshold
add expectedException.

- testGeeTesV4
This test use an url that seems to change the kind of captcha and are
skipeped until a new url are provided.